backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["MaksymPetyak <petyak.mi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
python-multipart = "^0.0.6"
requests = "^2.31.0"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
